refactor: rename visual_novel to visualization and exclude CHANGELOG
- Rename modules/visual_novel_widget.py → modules/visualization_widget.py
- Rename assets/visual_novel/ → assets/visualization/
- Update all code references from visual_novel to visualization
- Add CHANGELOG.md to .gitignore
- Update documentation and examples with new terminology

'visualization' is more accurate terminology for the avatar display widget.
